Alarms: Fallback to defualt when uri is invalid
This fixes an issue where the media is no longer in the media store
but since the URI uses the "content" scheme it is deemed a valid
URI even though the media is no longer available. We check if the
content is actually available and if not we will fallback to the
default ringtone.

Change-Id: I7b77265dfbf410681a245b3b020d838f3d1c7eba